Use sinatra to fake a Jira instance
The recent shift to JSONResource in Jiraby made spec testing more
difficult (the specific REST calls can't be easily stubbed because
RestClient::Resource instances are created on the fly with the `[]`
operator).

Stubbed JSON data is not the ideal way to test this stuff anyway;
something resembling a real Jira instance may be better in the long run.
This commit uses a MockJira sinatra app. Before running spec tests,
start the app with:

    $ rackup spec/mockapp/config.ru

(This will be automated in Rakefile eventually).