Make Issue an Entity, start implementing #save
This includes a hackish #from_json classmethod for initializing an Issue
with a JSONResource, the reason being that in order to save an Issue,
the headers from the Jira instance (in particular, the cookie that
indicates an active logged-in session) need to be included.

Also seen here are some crude methods for a kind of data-validation,
to prevent submitting the save (PUT) request with non-editable fields,
or with fields set to empty/nil. This is kind of a mess.